+++ include/robodash/impl/assets.hpp
@@ -5,7 +5,7 @@
  */
 
 #pragma once
-#include "apix.hpp"
+#include "robodash/apix.hpp"
 
 extern lv_font_t source_code_pro_16;
 extern lv_font_t montserrat_12;
