+++ include/robodash/impl/styles.hpp
@@ -5,7 +5,7 @@
  */
 
 #pragma once
-#include "apix.hpp"
+#include "robodash/apix.hpp"
 
 extern void _init_styles();
