+++ src/robodash/core.cpp
@@ -1,6 +1,5 @@
 #include "core.hpp"
 #include "apix.hpp"
-#include "filesystem.hpp"
 #include <map>
 #include <vector>
