+++ src/robodash/styles/initialize.cpp
@@ -1,4 +1,4 @@
-#include "robodash/styles.hpp"
+#include "robodash/apix.hpp"
 
 void _init_styles() {
 	_init_colors();
